float fix

+++ 3K.Application/Features/CekiIslemleri/DTOs/CekiSatiriDto.cs
@@ -7,7 +7,7 @@ namespace _3K.Application.Features.CekiIslemleri.DTOs
         public string? OlcuResmiPozNo { get; set; }
         public string BarkodNo { get; set; } = string.Empty;
         public string Aciklama { get; set; } = string.Empty;
-        public int IstenenAdet { get; set; }
+        public decimal IstenenAdet { get; set; }
         public string Birim { get; set; } = string.Empty;
         public string CekideGecenSandikNo { get; set; } = string.Empty;
         public string? FiiliSandikNo { get; set; }
@@ -15,8 +15,8 @@ namespace _3K.Application.Features.CekiIslemleri.DTOs
         public string Durum { get; set; } = string.Empty;
         public string? PaketleyenBasHarf { get; set; }
         public string? KontrolEdenBasHarf { get; set; }
-        public int KonulanAdet { get; set; }
-        public int EksikAdet { get; set; }
+        public decimal KonulanAdet { get; set; }
+        public decimal EksikAdet { get; set; }
     }
 }
