+++ 3K.Application/Features/GridIslemleri/Commands/GridManuelUrunEkleCommand.cs
@@ -10,7 +10,7 @@ namespace _3K.Application.Features.GridIslemleri.Commands
         public int ProjeId { get; set; }
         public string? BarkodNo { get; set; }
         public string Aciklama { get; set; } = string.Empty;
-        public int IstenenAdet { get; set; }
+        public decimal IstenenAdet { get; set; }
         public int? BirimId { get; set; }
         public string? EklemeNedeni { get; set; }
         public string SandikNo { get; set; } = string.Empty;
